+++ R/sim_design.R
@@ -1,4 +1,4 @@
-#' Simulate Data from Design
+#' Simulate data from design
 #'
 #' \code{sim_design()} generates a data table with a specified within and between design.
 #'
@@ -10,50 +10,42 @@
 #' @param r the correlations among the variables (can be a single number, vars\*vars matrix, vars\*vars vector, or a vars\*(vars-1)/2 vector)
 #' @param empirical logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance 
 #' @param long Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format
+#' @param dv the name of the dv for long plots (defaults to y)
+#' @param id the name of the id column (defaults to id)
 #' @param plot whether to show a plot of the design
 #' @param seed a single value, interpreted as an integer, or NULL (see set.seed)
 #' @param interactive whether to run the function interactively
+#' @param design a design list including within, between, n, mu, sd, r, dv, id
 #' 
 #' @return a tbl
 #' 
 #' @export
+#' @importFrom rlang := 
 #' 
 sim_design <- function(within = list(), between = list(), 
                        n = 100, mu = 0, sd = 1, r = 0, 
-                       empirical = FALSE, long = FALSE, 
-                       plot = FALSE, seed = NULL, interactive = FALSE) {
+                       empirical = FALSE, long = FALSE, dv = "y", id = "id",
+                       plot = TRUE, seed = NULL, interactive = FALSE, design = NULL) {
   # check the design is specified correctly
   if (interactive) {
     design <- interactive_design(plot = plot)
+  } else if (!is.null(design)) {
+    # double-check the entered design
+    design <- check_design(design = design, plot = plot)
+  } else if ("design" %in% class(within)) {
+    # design given as first argument: not ideal but handle it
+    design <- check_design(design = within, plot = plot)
   } else {
     design <- check_design(within = within, between = between, 
-                         n = n, mu = mu, sd = sd, r = r, plot = plot)
+                         n = n, mu = mu, sd = sd, r = r, 
+                         dv = dv, id = id, plot = plot)
   }
   
   # simulate the data
   sim_design_(design, empirical = empirical, long = long, seed = seed)
 }
 
-#' Fix name labels
-#' 
-#' Fixes if a factor list does not have named levels or has special characters in the names
-#' 
-#' @param x the list to fix
-#' 
-#' @return the fixed list
-#' @keywords internal
-#' 
-fix_name_labels <- function(x) {
-  if (is.null(names(x))) { names(x) <- x }
-  nm <- names(x)
-  # get rid of non-word characters and underscores because they mess up separate
-  names(x) <- gsub("(\\W|_)", ".", nm) 
-  x
-}
-
-#' Simulate Data from Design
-#'
-#' \code{sim_design_} generates a data table with a specified design
+#' Simulate data from design (internal)
 #'
 #' @param design A list of design parameters created by check_design()
 #' @param empirical logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance 
@@ -67,6 +59,10 @@ sim_design_ <- function(design, empirical = FALSE, long = FALSE, seed = NULL) {
   set.seed(seed)
   list2env(design, envir = environment())
   
+  # define columns
+  cells_w <- cell_combos(within, dv)
+  cells_b <- cell_combos(between, dv) 
+  
   # get factor names
   within_factors <- names(within)
   between_factors <- names(between)
@@ -76,28 +72,20 @@ sim_design_ <- function(design, empirical = FALSE, long = FALSE, seed = NULL) {
   if (length(within_factors) == 0)  within_factors  <- ".tmpvar." 
   
   # figure out number of subjects and their IDs
-  sub_n <- sum(cell_n[,1])
-  sub_id <- make_id(sub_n)
+  sub_n <- unlist(n) %>% sum()
   
   # simulate data for each between-cell
   for (cell in cells_b) {
-    if (length(within)) {
-      cell_vars <- rnorm_multi(
-        n = cell_n[cell,1], vars = length(cells_w), 
-        mu = cell_mu[cell,], sd = cell_sd[cell,], r = cell_r[[cell]], 
-        varnames = cells_w, empirical = empirical
-      ) %>%
-        dplyr::mutate("btwn" = cell)
-    } else {
-      # fully between design
-      sd2 <- cell_sd[cell,] %>% as.matrix() %>% as.vector()
-      val <- MASS::mvrnorm(n = cell_n[cell,1], 
-                           mu = cell_mu[cell,] %>% as.matrix() %>% as.vector(), 
-                           Sigma = sd2 %*% t(sd2), 
-                           empirical = empirical)
-      cell_vars <- data.frame("val" = val)  %>%
-        dplyr::mutate("btwn" = cell)
-    }
+    cell_vars <- rnorm_multi(
+      n = n[[cell]], 
+      vars = length(cells_w), 
+      mu = mu[[cell]] %>% unlist(), 
+      sd = sd[[cell]] %>% unlist(), 
+      r = r[[cell]], 
+      varnames = cells_w, 
+      empirical = empirical
+    ) %>%
+      dplyr::mutate("btwn" = cell)
     
     # add cell values to df
     if (cell == cells_b[1]) { 
@@ -108,13 +96,13 @@ sim_design_ <- function(design, empirical = FALSE, long = FALSE, seed = NULL) {
   }
   
   # set column order
-  col_order <- c("sub_id", between_factors, cells_w) %>%
+  col_order <- c(id, between_factors, cells_w) %>%
     setdiff(".tmpvar.")
   
   # create wide dataframe
   df_wide <- df %>%
     tidyr::separate("btwn", between_factors, sep = "_") %>%
-    dplyr::mutate("sub_id" = sub_id) %>%
+    dplyr::mutate(!!id := make_id(sub_n)) %>%
     dplyr::mutate_at(c(between_factors), ~as.factor(.)) %>%
     dplyr::select(tidyselect::one_of(col_order))
   
@@ -126,11 +114,11 @@ sim_design_ <- function(design, empirical = FALSE, long = FALSE, seed = NULL) {
   
   if (long == TRUE && length(within)) {
     # not necessary for fully between designs
-    col_order <- c("sub_id", between_factors, within_factors, "val") %>%
+    col_order <- c(id, between_factors, within_factors, dv) %>%
       setdiff(".tmpvar.")
     
     df_long <- df_wide %>%
-      tidyr::gather("w_in", "val", tidyselect::one_of(cells_w)) %>%
+      tidyr::gather("w_in", !!dv, tidyselect::one_of(cells_w)) %>%
       tidyr::separate("w_in", within_factors, sep = "_") %>%
       dplyr::select(tidyselect::one_of(col_order)) %>%
       dplyr::mutate_at(within_factors, ~as.factor(.))
@@ -141,8 +129,12 @@ sim_design_ <- function(design, empirical = FALSE, long = FALSE, seed = NULL) {
       df_long[[f]] <- factor(df_long[[f]], levels = names(within[[f]]))
     }
     
+    attr(df_long, "design") <- design
+    
     return(df_long)
   }
+  
+  attr(df_wide, "design") <- design
   
   df_wide
 }