+++ tests/testthat/test-check_design.R
@@ -6,19 +6,20 @@ test_that("2w", {
   between <- list()
   design <- check_design(within, between, n = 10, plot = FALSE)
   
-  cell_n <- data.frame(night = 10, day = 10, row.names = "val")
-  cell_mu <- data.frame(night = 0, day = 0, row.names = "val")
-  cell_sd <- data.frame(night = 1, day = 1, row.names = "val")
+  cell_n <- list(y = 10)
+  cell_mu <- list(y = list(night = 0, day = 0))
+  cell_sd <- list(y = list(night = 1, day = 1))
   
-  expect_equal(design$within, list(time = c(night = "night", day = "day")))
+  expect_equal(design$within, list(time = list(night = "night", day = "day")))
   expect_equal(design$between, list())
   
-  expect_equal(design$cell_n, cell_n)
-  expect_equal(design$cell_mu, cell_mu)
-  expect_equal(design$cell_sd, cell_sd)
+  expect_equal(design$n, cell_n)
+  expect_equal(design$mu, cell_mu)
+  expect_equal(design$sd, cell_sd)
+  expect_equal(design$dv, "y")
+  expect_equal(design$id, "id")
   
-  expect_equal(design$cells_w, c("night", "day"))
-  expect_equal(design$cells_b, "val")
+  expect_true("design" %in% class(design))
 })
 
 # 2b ----
@@ -27,19 +28,18 @@ test_that("2b", {
   between <- list(time = c("night", "day"))
   design <- check_design(within, between, n = 10, plot = FALSE)
   
-  cell_n <- data.frame(val = c(10, 10), row.names = c("night", "day"))
-  cell_mu <- data.frame(val = c(0, 0), row.names = c("night", "day"))
-  cell_sd <- data.frame(val = c(1, 1), row.names = c("night", "day"))
+  cell_n <- list(night = 10, day = 10)
+  cell_mu <- list(night = list(y=0), day = list(y=0))
+  cell_sd <- list(night = list(y=1), day = list(y=1))
   
   expect_equal(design$within, list())
-  expect_equal(design$between, list(time = c(night = "night", day = "day")))
+  expect_equal(design$between, list(time = list(night = "night", day = "day")))
   
-  expect_equal(design$cell_n, cell_n)
-  expect_equal(design$cell_mu, cell_mu)
-  expect_equal(design$cell_sd, cell_sd)
-  
-  expect_equal(design$cells_w, "val")
-  expect_equal(design$cells_b, c("night", "day"))
+  expect_equal(design$n, cell_n)
+  expect_equal(design$mu, cell_mu)
+  expect_equal(design$sd, cell_sd)
+  expect_equal(design$dv, "y")
+  expect_equal(design$id, "id")
 })
 
 # 2w*2b ----
@@ -48,22 +48,20 @@ test_that("2w*2b", {
   between <- list(pet = c("dog", "cat"))
   design <- check_design(within, between, n = 10, plot = FALSE)
 
-  cell_n <- data.frame(night = c(10,10), day = c(10,10), 
-                       row.names = c("dog", "cat"))
-  cell_mu <- data.frame(night = c(0,0), day = c(0,0), 
-                        row.names = c("dog", "cat"))
-  cell_sd <- data.frame(night = c(1,1), day = c(1,1), 
-                        row.names = c("dog", "cat"))
-
-  expect_equal(design$within, list(time = c(night = "night", day = "day")))
-  expect_equal(design$between, list(pet = c(dog = "dog", cat = "cat")))
+  cell_n <- list(dog = 10, cat = 10)
+  cell_mu <- list(dog = list(night = 0, day = 0),
+                  cat = list(night = 0, day = 0))
+  cell_sd <- list(dog = list(night = 1, day = 1),
+                  cat = list(night = 1, day = 1))
   
-  expect_equal(design$cell_n, cell_n)
-  expect_equal(design$cell_mu, cell_mu)
-  expect_equal(design$cell_sd, cell_sd)
+  expect_equal(design$within, list(time = list(night = "night", day = "day")))
+  expect_equal(design$between, list(pet = list(dog = "dog", cat = "cat")))
   
-  expect_equal(design$cells_w, c("night", "day"))
-  expect_equal(design$cells_b, c("dog", "cat"))
+  expect_equal(design$n, cell_n)
+  expect_equal(design$mu, cell_mu)
+  expect_equal(design$sd, cell_sd)
+  expect_equal(design$dv, "y")
+  expect_equal(design$id, "id")
 })
 
 # 2w*2w*2b*2b ----
@@ -81,15 +79,21 @@ test_that("2w*2w*2b*2b", {
   
   cells_w <- c("day_A", "night_A", "day_B", "night_B")
   cells_b <- c("dog_old", "cat_old", "dog_young", "cat_young")
-  cell_n <- matrix(rep(100,16), 4, dimnames = list(cells_b, cells_w)) %>% as.data.frame()
-  cell_mu <- matrix(rep(0,16), 4, dimnames = list(cells_b, cells_w)) %>% as.data.frame()
-  cell_sd <- matrix(rep(1,16), 4, dimnames = list(cells_b, cells_w)) %>% as.data.frame()
+  cell_n <- list(dog_old = 100, cat_old = 100, dog_young = 100, cat_young = 100)
+  cell_mu <- list(dog_old = list(day_A = 0, night_A = 0, day_B = 0, night_B = 0),
+                  cat_old = list(day_A = 0, night_A = 0, day_B = 0, night_B = 0),
+                  dog_young = list(day_A = 0, night_A = 0, day_B = 0, night_B = 0),
+                  cat_young = list(day_A = 0, night_A = 0, day_B = 0, night_B = 0))
+  cell_sd <- list(dog_old = list(day_A = 1, night_A = 1, day_B = 1, night_B = 1),
+                  cat_old = list(day_A = 1, night_A = 1, day_B = 1, night_B = 1),
+                  dog_young = list(day_A = 1, night_A = 1, day_B = 1, night_B = 1),
+                  cat_young = list(day_A = 1, night_A = 1, day_B = 1, night_B = 1))
   
-  expect_equal(design$cells_w, cells_w)
-  expect_equal(design$cells_b, cells_b)
-  expect_equal(design$cell_n, cell_n)
-  expect_equal(design$cell_mu, cell_mu)
-  expect_equal(design$cell_sd, cell_sd)
+  expect_equal(design$n, cell_n)
+  expect_equal(design$mu, cell_mu)
+  expect_equal(design$sd, cell_sd)
+  expect_equal(design$dv, "y")
+  expect_equal(design$id, "id")
 })
 
 # design spec ----
@@ -116,13 +120,16 @@ test_that("design spec", {
     "B1" = .2,
     "B2" = .5
   )
+  dv <- "dv"
+  id <- "sub_id"
   
-  design <- check_design(within, between, n, mu, sd, r, plot = FALSE)
+  design <- check_design(within, between, n, mu, sd, r, dv, id, plot = FALSE)
   
-  design_elements <- c("within", "between", "cells_w", "cells_b", 
-                       "cell_n", "cell_mu", "cell_sd", "cell_r")
+  design_elements <- c("within", "between", "dv", "id", "n", "mu", "sd", "r")
   
   expect_equal(names(design), design_elements)
+  expect_equal(design$dv, dv)
+  expect_equal(design$id, id)
 })
 
 # anon factors ----
@@ -130,13 +137,13 @@ test_that("anon factors", {
   design <- check_design(c(2, 4), c(2, 2), plot = FALSE)
   
   w <- list(
-    A = c(A1="A1", A2="A2"),
-    B = c(B1="B1", B2="B2", B3="B3", B4="B4")
+    A = list(A1="A1", A2="A2"),
+    B = list(B1="B1", B2="B2", B3="B3", B4="B4")
   )
   
   b <- list(
-    C = c(C1="C1",C2="C2"),
-    D = c(D1="D1", D2="D2")
+    C = list(C1="C1",C2="C2"),
+    D = list(D1="D1", D2="D2")
   )
   
   expect_equal(design$within, w)
@@ -164,3 +171,4 @@ test_that("make_id", {
   expect_equal(make_id(2:4), c("S2", "S3", "S4"))
   expect_equal(make_id(100:200)[[1]], "S100")
 })
+