+++ tests/testthat/test-interactive_design.R
@@ -1,18 +1,30 @@
 context("test-interactive_design")
 
-# test_that("test", {
-#   skip_on_cran()
-#   skip_on_travis()
-#   
-#   des <- interactive_design()
-#   
-#   #des <- check_design(d$within, d$between, d$n, d$mu, d$sd, d$r, plot = FALSE)
-#   des2 <- check_design(2, 2, plot = FALSE)
-#   
-#   expect_equal(des$within, des2$within)
-#   expect_equal(des$between, des2$between)
-#   expect_equal(des$cell_n, des2$cell_n)
-#   expect_equal(des$cell_mu, des2$cell_mu)
-#   expect_equal(des$cell_sd, des2$cell_sd)
-#   expect_equal(des$cell_r, des2$cell_r)
-# })
+test_that("test", {
+  skip("interactive is too annoying")
+
+  within <- list(A = c("A1", "A2", "A3"))
+  between <- list(B = c("B1", "B2"))
+  n <- 10
+  mu <- list(
+    B1 = 1:3,
+    B2 = 4:6
+  )
+  sd <- 1
+  r <- list(
+    B1 = .5,
+    B2 = c(.4, .5, .6)
+  )
+  des2 <- check_design(within, between, n, mu, sd, r, plot = TRUE)
+  
+  des <- interactive_design()
+  
+  expect_equal(des$within,  des2$within)
+  expect_equal(des$between, des2$between)
+  expect_equal(des$id, des2$id)
+  expect_equal(des$dv, des2$dv)
+  expect_equal(des$n,  des2$n)
+  expect_equal(des$mu, des2$mu)
+  expect_equal(des$sd, des2$sd)
+  expect_equal(des$r,  des2$r)
+})